+++ sql/mis/sd_datransact.sql
@@ -20,3 +20,25 @@ CREATE TABLE IF NOT EXISTS tab1 (
 );
 CREATE INDEX idx ON tab1 (report_date);
 
+
+INSERT INTO tab1 
+SELECT 
+    account_id::UINTEGER,
+    report_date::DATE,
+    version::TIMESTAMP,
+    strptime(left(hour_beginning, 25), '%Y-%m-%dT%H:%M:%S%z')::TIMESTAMPTZ,
+    transaction_number::UINTEGER,
+    reference_id::VARCHAR,
+    transaction_type::ENUM ('IBM'),
+    other_party::UINTEGER,
+    settlement_location_id::UINTEGER,
+    location_name::VARCHAR,
+    location_type::ENUM ('HUB', 'LOAD ZONE', 'NETWORK NODE', 'DRR AGGREGATION ZONE'),
+    amount::DECIMAL(9,4),
+    impacts_marginal_loss_revenue_allocation::BOOLEAN,
+    subaccount_id::VARCHAR
+FROM read_csv(
+    'C:/Temp/MIS/SD_DATRANSACT/tmp/tab1_*.CSV', 
+    header = true, 
+    timestampformat = '%Y-%m-%dT%H:%M:%SZ'
+);